Introduce `PINFuturePending` for handling delegate patterns
This commit introduces `PINFuturePending`, a wrapper on top of `PINFuture`
that could be used to handle cases where we expect the future to be resolved
later in some other part of the code, e.g. 3rd party delegate method that
we have little control over when it will get called and resolve or reject the
future based on the result. The benefit of this is that we'll be able to chain
success/failure methods right away to the future without having to wait for it
to resolve or get rejected.